requirements.txt
# Core ML and data
torch>=2.2.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.12.0

# Retrieval and indexing
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
httpx>=0.27.0

# Streaming
kafka-python>=2.0.2

# Online feature store
redis>=5.0.0

# Evaluation and reporting
tabulate>=0.9.0

# Parquet support
pyarrow>=15.0.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0
requests>=2.31.0
python-multipart>=0.0.9
